+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ v0.5.6 (XXXX-XX-XX)
 * updated Daemon constructor to support Windows out of the box
 * fixed `copy_from_host` and `copy_to_host` in `FileSystem` to support
   Windows hosts.
+* removed use of `signal` module inside implementation of `Popen` to
+  improve support for Windows host.
 
 
 v0.5.5 (2020-07-08)